rewite SetByRules based on the review comments:
This will not work as a resource function because the UpValues are not defined for ResourceFunction, which is the actual head for WFR items. So the parts that use UpValues will need to be removed. Try generating a preview from this notebook to see how it works as a ResourceFunction.